refactor: based on PR feedbacks
