+++ src/components/SideNav.tsx
@@ -8,7 +8,12 @@ import List from "@mui/material/List";
 import MenuIcon from "@mui/icons-material/Menu";
 import Typography from "@mui/material/Typography";
 
-export default function SideNav({ children }: any) {
+// An interface that defines object parameters
+interface IProps {
+    children?: React.ReactNode;
+};
+
+export default function SideNav({ children }:IProps) {
     const [mobileOpen, setMobileOpen] = React.useState(false);
     const handleDrawerToggle = () => {
         setMobileOpen(!mobileOpen);
@@ -62,6 +67,7 @@ const style = {
             boxSizing: "border-box",
             width: 180,
         },
+        zIndex: "1250!important"
     },
     margin: {
         iconBtn: { ml: 0.5 },
