+++ src/theme/ThemeHandler.tsx
@@ -3,8 +3,13 @@ import CssBaseline from "@mui/material/CssBaseline";
 import { createTheme, ThemeProvider } from "@mui/material/styles";
 import { ColorModeContext } from "../context/appContext";
 import { baseTheme } from "./Themes";
+// An interface that defines object parameters
+interface IProps {
+    children?: React.ReactNode;
+};
+// getting the localstorage item
 const savedMode = window.localStorage.getItem("colorMode") as "light" | "dark";
-export default function ThemeHandler({ children }: any) {
+export default function ThemeHandler({ children }:IProps) {
     // Use localStorage.getItem save the toggle color state
     // When the page is refreshed, it remains the current toggled mode
     const [mode, setMode] = useState(savedMode || "dark");
