+++ src/theme/__tests__/Themes.test.tsx
@@ -1,4 +1,5 @@
-import { baseTheme } from "../components/Themes";
+import { baseTheme } from "../Themes";
+
 test("theme is used", () => {
     const wrapper = baseTheme("light");
     expect(wrapper.palette.mode).toBe("light");
